+++ changelog.md
@@ -1,5 +1,6 @@
 # 1.4.1
-- Added garage preview support (toggleable)
+- Added garage preview support
+- Added profile icon mod support and compatibility
 
 # 1.4.0
 - Fixed the flashPlayer wich was making the icon white permanently